Release 1.2.1

+++ modules/users/controllers/delete.js
@@ -1,5 +1,6 @@
 'use strict';
 
+const policy = require('../user_policy');
 
 /**
  *
@@ -14,19 +15,17 @@
 
 module.exports = (db) => async function (req, res) {
     if (isNaN(parseInt(req.params.userId, 10))) {
-        throw res.boom.badRequest();
+        throw res.boom.badRequest('User ID must be an integer');
     }
     const userId = parseInt(req.params.userId, 10);
 
-    try {
-        await db.UserGroup.destroy({where: {userId}});
-        await db.User.destroy({where: {id: userId}});
-
-        return res.status(204).send();
-    } catch (err) {
-        if (err instanceof db.Sequelize.ValidationError) {
-            throw res.boom.badRequest();
-        }
-        throw err;
+    let authorized = policy.filterDelete(db, req.params.userId, req.session.auth);
+    if(!authorized) {
+        return res.boom.unauthorized();
     }
+
+    await db.UserGroup.destroy({where: {userId}});
+    await db.User.destroy({where: {id: userId}});
+
+    return res.status(204).send();
 };
